+++ src/player.h
@@ -344,8 +344,8 @@ class Player final : public Creature, public Cylinder
 
 		bool canOpenCorpse(uint32_t ownerId) const;
 
-		void addStorageValue(const uint32_t key, const int32_t value, const bool isLogin = false);
-		bool getStorageValue(const uint32_t key, int32_t& value) const;
+		void addStorageValue(const uint32_t key, const int64_t value, const bool isLogin = false);
+		bool getStorageValue(const uint32_t key, int64_t& value) const;
 		void genReservedStorageRange();
 
 		void setGroup(Group* newGroup) {
@@ -1201,7 +1201,7 @@ class Player final : public Creature, public Cylinder
 		std::map<uint8_t, OpenContainer> openContainers;
 		std::map<uint32_t, DepotLocker*> depotLockerMap;
 		std::map<uint32_t, DepotChest*> depotChests;
-		std::map<uint32_t, int32_t> storageMap;
+		std::map<uint32_t, int64_t> storageMap;
 
 		std::vector<OutfitEntry> outfits;
 		GuildWarVector guildWarVector;
